requirements.txt
fastapi
uvicorn
scikit-learn
logtail-python
